Pull container usage out to app
