+++ components/experimental/diceroller-standalone/src/app/loader.ts
@@ -15,11 +15,7 @@ import {
 import { Container } from "@fluidframework/container-loader";
 import { IUser } from "@fluidframework/protocol-definitions";
 import { RouterliciousDocumentServiceFactory, DefaultErrorTracking } from "@fluidframework/routerlicious-driver";
-import { HTMLViewAdapter } from "@fluidframework/view-adapters";
-import { IComponentMountableView } from "@fluidframework/view-interfaces";
 import { extractPackageIdentifierDetails } from "@fluidframework/web-code-loader";
-import { IComponent } from "@fluidframework/component-core-interfaces";
-import { RequestParser } from "@fluidframework/runtime-utils";
 import { getRandomName } from "@fluidframework/server-services-client";
 // eslint-disable-next-line import/no-internal-modules
 import uuid from "uuid/v4";
@@ -62,8 +58,7 @@ export async function start(
     packageJson: IFluidPackage,
     fluidModule: IFluidModule,
     port: number,
-    div: HTMLDivElement,
-): Promise<void> {
+): Promise<Container> {
     const documentServiceFactory = new RouterliciousDocumentServiceFactory(
         false,
         new DefaultErrorTracking(),
@@ -101,58 +96,9 @@ export async function start(
         hostConf,
         [packageSeed],
     );
-    let container: Container;
 
-    container = await baseHost.initializeContainer(
+    return await baseHost.initializeContainer(
         url,
         codeDetails,
     );
-
-    // Needs updating if the doc id is in the hash
-    const reqParser = new RequestParser({ url });
-    const componentUrl = `/${reqParser.createSubRequest(3)!.url}`;
-
-    await getComponentAndRender(container, componentUrl, div);
-    // Handle the code upgrade scenario (which fires contextChanged)
-    container.on("contextChanged", () => {
-        getComponentAndRender(container, componentUrl, div).catch(() => { });
-    });
-}
-
-async function getComponentAndRender(container: Container, url: string, div: HTMLDivElement) {
-    const response = await container.request({
-        headers: {
-            mountableView: true,
-        },
-        url,
-    });
-
-    if (response.status !== 200 ||
-        !(
-            response.mimeType === "fluid/component" ||
-            response.mimeType === "prague/component"
-        )) {
-        return false;
-    }
-
-    const component = response.value as IComponent;
-    if (component === undefined) {
-        return;
-    }
-
-    // We should be retaining a reference to mountableView long-term, so we can call unmount() on it to correctly
-    // remove it from the DOM if needed.
-    const mountableView: IComponentMountableView | undefined = component.IComponentMountableView;
-    if (mountableView !== undefined) {
-        mountableView.mount(div);
-        return;
-    }
-
-    // If we don't get a mountable view back, we can still try to use a view adapter.  This won't always work (e.g.
-    // if the response is a React-based component using hooks) and is not the preferred path, but sometimes it
-    // can work.
-    console.warn(`Container returned a non-IComponentMountableView.  This can cause errors when mounting components `
-        + `with React hooks across bundle boundaries.  URL: ${url}`);
-    const view = new HTMLViewAdapter(component);
-    view.render(div, { display: "block" });
 }
